AST

Description

This module contains functions for the construction of abstract syntax tree's. The abstract syntax tree is defined as a lookup table with characters as keys and StackNode's (defined in the Stack module) as functions.

Documentation

getOperationsTillFunction :: [String] -> [String]

Get all lines until a new function definition (?) is found.

clusterFunctions :: [String] -> [(Char, [String])]

Cluster functions into groups with a function name (names are always just one character)

createFlowNode :: String -> MaybeError FlowNode

Creates a FlowNode based on a single string. Depending on the content of the string it will either be a AssignmentNode or a ConditionNode.

createConditionNodeAbstract :: String -> Char -> String -> MaybeError FlowNode

Parses two strings to ExpressionNode and create a ConditionNode.

getExpression :: String -> MaybeError ExpressionNode

Read a string, check what kind of expression node should be used and create the expression node.

getExpressionNumbers :: String -> MaybeError ExpressionNode

Maybe get an expression node when you are sure a string only contains numbers

getExpressionVariable :: String -> MaybeError ExpressionNode

Maybe get an expression node when you are sure a string only contains a character that can be used as a variable

getExpressionWithMath :: String -> Int -> MaybeError ExpressionNode

Maybe get an expression node when you are sure a string contains a mathematical symbol

validateExpressions :: MaybeError ExpressionNode -> Operator -> MaybeError ExpressionNode -> MaybeError ExpressionNode

Check if two expressions are valid and combine them with an operator

allNumbers :: String -> Bool

Check if a list only contains digits.

createFunctionNodes :: [(Char, [String])] -> MaybeError [(Char, StackNode)]

Converts a given list with characters and lists of strings to a stack.

createFunctionAST :: [String] -> MaybeError StackNode

Wrapper around _createFunctionAST to convert [FlowNode] to StackNode.

_createFunctionAST :: [String] -> MaybeError [FlowNode]

Create an AST(abstract syntax tree) for a list of strings (a function).

createAST :: [String] -> MaybeError [(Char, StackNode)]

Create AST's for all functions and return stack containing these functions.