Reader

Description

This module is used for reading a file.

Documentation

readOneFile :: String -> MaybeError (IO String)

Read a given file and return the filecontent as an IO String. If the given file does not exist the return value is an Error.

cleanInput :: String -> MaybeError [String]

Remove whitespaces, split the string on newlines and check if the first line contains a questionmark. The first line must be a questionmark otherwise there are operations outside functions. Error is returned if the first line does not contain a questionmark.

checkFirstLine :: [String] -> MaybeError [String]

Check if the first character of the first line is a questionmark, otherwise return Error.

isValidFilename :: String -> Bool

Check if a given string is a valid filename, this is only the case when the filename consists of one character and the .one extension.

removeWhitespace :: String -> String

remove tabs, spaces or carriage return characters

removeEmptyLines :: [String] -> [String]

remove empty lines, or lines that start with #, from a list of strings