Runner

Description

This module is used for running an abstract syntax tree (defined in the AST module).

Documentation

run :: [(Char, StackNode)] -> (MaybeError Int, [(Char, StackNode)])

Run a given stack starting at function a. Return the return value and the resulting stack. The return value can be an error, then the stack can be used alongside with the error for debugging.

executeFunction :: [FlowNode] -> Bool -> Int -> [(Char, StackNode)] -> (MaybeError Int, [(Char, StackNode)])

Execute a list of FlowNode's. If the bool argument is False the next statement will be ignored. This happens after conditionals. The int is the last used value and can be the return value of the function. And the last argument is the stack. The return tuple contains the result and the resulting stack.

executeFlowNode :: FlowNode -> [(Char, StackNode)] -> (Bool, MaybeError Int, [(Char, StackNode)])

Execute a single FlowNode. |Return whether the next statement must be skipped, the resulting int and the resulting stack.

conditionHelper :: ExpressionNode -> ExpressionNode -> (Int -> Int -> Bool) -> [(Char, StackNode)] -> (Bool, MaybeError Int, [(Char, StackNode)])

Uses two expressions a function and a stack to test a condition and return the result.

getExpressionResult :: ExpressionNode -> [(Char, StackNode)] -> (MaybeError Int, [(Char, StackNode)])

Gets the result of an ExpressionNode, this is dependent on the given stack.

operationHelper :: ExpressionNode -> ExpressionNode -> (Int -> Int -> Int) -> [(Char, StackNode)] -> (MaybeError Int, [(Char, StackNode)])

An abstraction layer for handeling mathmatical operations.

divide :: Int -> Int -> Int

Integer division.

getStackResult :: StackNode -> [(Char, StackNode)] -> (MaybeError Int, [(Char, StackNode)])

Gets a value from the stack. This can either be a simple integer or the result of a function.