Stack

Description

This module is used for creating a stack that can bee searched and edited.

Documentation

type Operator = Char

Used to define parts of conditions and expressions.

data FlowNode

These are the "steps" that will be taken when executing the program.

Constructors

ConditionNode ExpressionNode Operator ExpressionNode

AssignmentNode ExpressionNode ExpressionNode

data StackNode

These are objects that ar defined on the stack, either a value or a StartNode, which is a list of flownodes which is a function.

Constructors

ValueNode Int

StartNode [FlowNode]

data ExpressionNode

Expressions are built using expressionNodes. They can either be numbers, variables or functions defined on the stack or subexpressions.

Constructors

IntNode Int

OperationNode ExpressionNode Operator ExpressionNode

StackVariableNode Char

getFlowNodes :: StackNode -> [FlowNode]

Gets the list of FlowNode's.

isFunction :: StackNode -> Bool

Check if function

getFromStack :: Char -> [(Char, StackNode)] -> MaybeError StackNode

Get a StackNode from stack. When the stack gets bigger this function becomes slower as simple lists are used, not a binary tree.

setStack :: Char -> StackNode -> [(Char, StackNode)] -> MaybeError [(Char, StackNode)]

Set a value on the stack. When the stack gets bigger this function becomes slower as simple lists are used, not a binary tree. |This wraps _stackSet with a check for illegal characters.

_stackSet :: Char -> StackNode -> [(Char, StackNode)] -> [(Char, StackNode)]

Set a value on the stack. When the stack gets bigger this function becomes slower as simple lists are used, not a binary tree.

prettyStrStack :: [(Char, StackNode)] -> String

Get a string that represents the stack nicely.

newStack :: [(Char, StackNode)]

Create empty stack