Tools

Description

This module contains more general functions and is used in almost all other modules.

Documentation

data MaybeError a

Data that contains either a value or an error. This is used heavily as backtrace for errors.

Constructors

NotError a

Error String

isValue :: MaybeError a -> Bool

Check if a MaybeError is a value and not an error.

getValue :: MaybeError a -> a

Get the value from a MaybeError. Note that this will fail if the MaybeError was an Error.

isError :: MaybeError a -> Bool

Check if a MaybeError is an error and not a value.

getError :: MaybeError a -> String

Get the error from a MaybeError. Note that this will fail if the MaybeError was a Value.

conditionalRemove :: [a] -> (a -> Bool) -> [a]

Removes elements from list that return true on given function.

_splitOn :: [b] -> (b -> Bool) -> [b] -> [b] -> [[b]]

This function splits a list using a given function. The third and fourth parameters should be empty lists.

splitString :: String -> Char -> [String]

Wrapper around _splitOn for strings.

isIllegalVariableCharacter :: Char -> Bool

Check if a character is invalid. This is the case when it is either a comparison operator (=) a mathematical operator (+-*/) a digit (0..9) an assignment operator (~) or a comment identifier (#).

isMathmaticalOperation :: Char -> Bool

Check if a character is a mathematical operator (+-*/).

fst3 :: (a, b, c) -> a

Get first element in tuple.

snd3 :: (a, b, c) -> b

Get second element in tuple.

trd3 :: (a, b, c) -> c

Get third element in tuple.

boolToInt :: Bool -> Int

Convert bool too int (True = 1, False = 0).

getFirstArg :: [a] -> MaybeError a

Get first element from list or return error

lastArgIsDebug :: [String] -> Bool

Check if the last element of a list of strings is "debug"